fix: resolve all frontend ESLint warnings
Fix 7 react-refresh/only-export-components warnings by moving
non-component exports into existing utils/store modules, and 15
react-hooks/exhaustive-deps warnings by completing dependency arrays
with stable refs/setters and destructuring stable close callbacks
from anchored panel menus. No rule disables or config changes.
