+++ frontend/src/components/chat/model-selector/ModelSelector.tsx
@@ -9,7 +9,8 @@ import { useAuthStore } from '@/store/authStore';
 import { useModelStore } from '@/store/modelStore';
 import { useModelsQuery } from '@/hooks/queries/useModelQueries';
 import { useIsSplitMode } from '@/hooks/useIsSplitMode';
-import { AGENT_ICONS, ProviderIcon } from '@/components/ui/icons/ProviderIcon';
+import { ProviderIcon } from '@/components/ui/icons/ProviderIcon';
+import { AGENT_ICONS } from '@/components/ui/icons/providerIcons';
 import { AgentFilterChips } from '@/components/chat/model-selector/AgentFilterChips';
 import { formatNumberCompact } from '@/utils/format';
 import type { AgentKind, Model } from '@/types/chat.types';
