+++ frontend/src/components/layout/ChatTabs/ChatTabs.tsx
@@ -2,7 +2,8 @@ import { useCallback, useEffect, useMemo } from 'react';
 import { useMatch, useNavigate } from 'react-router-dom';
 import { X } from 'lucide-react';
 import { AsciiSpinner } from '@/components/ui/AsciiSpinner/AsciiSpinner';
-import { ChatStatusDot, type ChatStatusTone } from '@/components/ui/ChatStatusDot/ChatStatusDot';
+import { ChatStatusDot } from '@/components/ui/ChatStatusDot/ChatStatusDot';
+import type { ChatStatusTone } from '@/utils/message';
 import { Button } from '@/components/ui/primitives/Button/Button';
 import { FloatingTooltip } from '@/components/ui/FloatingTooltip/FloatingTooltip';
 import { ProviderIcon } from '@/components/ui/icons/ProviderIcon';
