+++ frontend/src/components/layout/Sidebar/SubThreadList.tsx
@@ -3,11 +3,8 @@ import { MoreHorizontal } from 'lucide-react';
 import { useSubThreadsQuery } from '@/hooks/queries/useChatQueries';
 import { useChatAgentKind } from '@/hooks/useChatAgentKind';
 import { AsciiSpinner } from '@/components/ui/AsciiSpinner/AsciiSpinner';
-import {
-  ChatStatusDot,
-  chatStatusTone,
-  type ChatStatusTone,
-} from '@/components/ui/ChatStatusDot/ChatStatusDot';
+import { ChatStatusDot } from '@/components/ui/ChatStatusDot/ChatStatusDot';
+import { chatStatusTone, type ChatStatusTone } from '@/utils/message';
 import { Button } from '@/components/ui/primitives/Button/Button';
 import { FloatingTooltip } from '@/components/ui/FloatingTooltip/FloatingTooltip';
 import { ProviderIcon } from '@/components/ui/icons/ProviderIcon';
