+++ frontend/src/components/ui/icons/ProviderIcon.tsx
@@ -1,20 +1,6 @@
-import type { ComponentType, SVGProps } from 'react';
-import { ClaudeIcon } from './ClaudeIcon';
-import { CodexIcon } from './CodexIcon';
-import { CopilotIcon } from './CopilotIcon';
-import { CursorIcon } from './CursorIcon';
-import { GrokIcon } from './GrokIcon';
-import { OpencodeIcon } from './OpencodeIcon';
+import type { SVGProps } from 'react';
 import type { AgentKind } from '@/types/chat.types';
-
-export const AGENT_ICONS: Record<AgentKind, ComponentType<SVGProps<SVGSVGElement>>> = {
-  claude: ClaudeIcon,
-  codex: CodexIcon,
-  copilot: CopilotIcon,
-  cursor: CursorIcon,
-  grok: GrokIcon,
-  opencode: OpencodeIcon,
-};
+import { AGENT_ICONS } from './providerIcons';
 
 interface ProviderIconProps extends SVGProps<SVGSVGElement> {
   agentKind: AgentKind;
