+++ frontend/src/hooks/useInputEnhance.ts
@@ -25,7 +25,7 @@ export function useInputEnhance({
         textarea.setSelectionRange(length, length);
       }, 0);
     }
-  }, []);
+  }, [textareaRef]);
 
   const enhancePromptMutation = useEnhancePromptMutation({
     onSuccess: (enhancedPrompt) => {
@@ -39,7 +39,7 @@ export function useInputEnhance({
   const handleEnhancePrompt = useCallback(() => {
     if (!hasMessage || isEnhancing) return;
     enhancePromptMutation.mutate({ prompt: messageRef.current.trim(), modelId: selectedModelId });
-  }, [hasMessage, isEnhancing, selectedModelId, enhancePromptMutation]);
+  }, [hasMessage, isEnhancing, selectedModelId, enhancePromptMutation, messageRef]);
 
   return {
     isEnhancing,
